requirements.txt
deepgram-sdk>=2.12.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
websockets>=11.0.3
pydantic>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.1
numpy>=1.24.0
pyaudio>=0.2.13
python-multipart>=0.0.6
